register page alert bug fixed usefull comment added todo list added
